+++ src/drive/styles/dropzone.styl
@@ -29,18 +29,17 @@
     margin-bottom 1rem
     border-radius 100%
     border none
-    background-color dodgerBlue
-    background-image embedurl('../assets/icons/icon-claudy.svg')
-    background-repeat no-repeat
-    background-size 2.5em
-    background-position .5em
     box-shadow 0 1px 3px 0 rgba(50, 54, 63, 0.19), 0 6px 18px 0 rgba(50, 54, 63, 0.39)
     animation-name pulse
     animation-duration 2s
     animation-iteration-count infinite
     cursor pointer
     outline 0
-
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background-color dodgerBlue
+    
 .fil-dropzone-teaser-content
     background-color dodgerBlue
     border-radius 2px
